feat: mvp signup
