--- apps/web/src/components/Shared/Auth/Signup.tsx
@@ -1,190 +1,0 @@
-import type {
-  LastLoggedInProfileRequest,
-  Profile,
-  ProfileManagersRequest
-} from '@hey/lens';
-import type { Dispatch, FC, SetStateAction } from 'react';
-
-import SwitchNetwork from '@components/Shared/SwitchNetwork';
-import { ArrowRightCircleIcon, KeyIcon } from '@heroicons/react/24/outline';
-import { XCircleIcon } from '@heroicons/react/24/solid';
-import { Errors } from '@hey/data/errors';
-import { AUTH } from '@hey/data/tracking';
-import {
-  useAuthenticateMutation,
-  useChallengeLazyQuery,
-  useProfilesManagedQuery
-} from '@hey/lens';
-import { Button, Card, Spinner } from '@hey/ui';
-import errorToast from '@lib/errorToast';
-import { Leafwatch } from '@lib/leafwatch';
-import { useState } from 'react';
-import toast from 'react-hot-toast';
-import { CHAIN } from 'src/constants';
-import { signIn } from 'src/store/persisted/useAuthStore';
-import { useAccount, useChainId, useDisconnect, useSignMessage } from 'wagmi';
-
-import UserProfile from '../UserProfile';
-import WalletSelector from './WalletSelector';
-
-interface SignupProps {
-  setHasConnected?: Dispatch<SetStateAction<boolean>>;
-}
-
-const Signup: FC<SignupProps> = ({ setHasConnected }) => {
-  const [isLoading, setIsLoading] = useState(false);
-  const [loggingInProfileId, setLoggingInProfileId] = useState<null | string>(
-    null
-  );
-
-  const onError = (error: any) => {
-    setIsLoading(false);
-    errorToast(error);
-  };
-
-  const chain = useChainId();
-  const { disconnect } = useDisconnect();
-  const { address, connector: activeConnector } = useAccount();
-  const { signMessageAsync } = useSignMessage({ mutation: { onError } });
-  const [loadChallenge, { error: errorChallenge }] = useChallengeLazyQuery({
-    fetchPolicy: 'no-cache'
-  });
-  const [authenticate, { error: errorAuthenticate }] =
-    useAuthenticateMutation();
-  const request: LastLoggedInProfileRequest | ProfileManagersRequest = {
-    for: address
-  };
-  const { data: profilesManaged, loading: profilesManagedLoading } =
-    useProfilesManagedQuery({
-      skip: !address,
-      variables: {
-        lastLoggedInProfileRequest: request,
-        profilesManagedRequest: request
-      }
-    });
-
-  const handleSign = async (id?: string) => {
-    try {
-      setLoggingInProfileId(id || null);
-      setIsLoading(true);
-      // Get challenge
-      const challenge = await loadChallenge({
-        variables: { request: { ...(id && { for: id }), signedBy: address } }
-      });
-
-      if (!challenge?.data?.challenge?.text) {
-        return toast.error(Errors.SomethingWentWrong);
-      }
-
-      // Get signature
-      const signature = await signMessageAsync({
-        message: challenge?.data?.challenge?.text
-      });
-
-      // Auth user and set cookies
-      const auth = await authenticate({
-        variables: { request: { id: challenge.data.challenge.id, signature } }
-      });
-      const accessToken = auth.data?.authenticate.accessToken;
-      const refreshToken = auth.data?.authenticate.refreshToken;
-      signIn({ accessToken, refreshToken });
-      Leafwatch.track(AUTH.SIWL);
-      location.reload();
-    } catch {}
-  };
-
-  const allProfiles = profilesManaged?.profilesManaged.items || [];
-  const lastLogin = profilesManaged?.lastLoggedInProfile;
-
-  const remainingProfiles = lastLogin
-    ? allProfiles.filter((profile) => profile.id !== lastLogin.id)
-    : allProfiles;
-
-  const profiles = lastLogin
-    ? [lastLogin, ...remainingProfiles]
-    : remainingProfiles;
-
-  return activeConnector?.id ? (
-    <div className="space-y-3">
-      <div className="space-y-2.5">
-        {chain === CHAIN.id ? (
-          profilesManagedLoading ? (
-            <Card className="w-full dark:divide-gray-700" forceRounded>
-              <div className="space-y-2 p-4 text-center text-sm font-bold">
-                <Spinner className="mx-auto" size="sm" />
-                <div>Loading profiles managed by you...</div>
-              </div>
-            </Card>
-          ) : profiles.length > 0 ? (
-            <Card className="w-full dark:divide-gray-700" forceRounded>
-              {profiles.map((profile) => (
-                <div
-                  className="flex items-center justify-between p-3"
-                  key={profile.id}
-                >
-                  <UserProfile
-                    linkToProfile={false}
-                    profile={profile as Profile}
-                    showUserPreview={false}
-                  />
-                  <Button
-                    disabled={isLoading && loggingInProfileId === profile.id}
-                    icon={
-                      isLoading && loggingInProfileId === profile.id ? (
-                        <Spinner size="xs" />
-                      ) : (
-                        <ArrowRightCircleIcon className="size-4" />
-                      )
-                    }
-                    onClick={() => handleSign(profile.id)}
-                  >
-                    Login
-                  </Button>
-                </div>
-              ))}
-            </Card>
-          ) : (
-            <div>
-              <Button
-                disabled={isLoading}
-                icon={
-                  isLoading ? (
-                    <Spinner size="xs" />
-                  ) : (
-                    <ArrowRightCircleIcon className="size-4" />
-                  )
-                }
-                onClick={() => handleSign()}
-              >
-                Sign in with Lens
-              </Button>
-            </div>
-          )
-        ) : (
-          <SwitchNetwork toChainId={CHAIN.id} />
-        )}
-        <button
-          className="flex items-center space-x-1 text-sm underline"
-          onClick={() => {
-            disconnect?.();
-            Leafwatch.track(AUTH.CHANGE_WALLET);
-          }}
-          type="reset"
-        >
-          <KeyIcon className="size-4" />
-          <div>Change wallet</div>
-        </button>
-      </div>
-      {errorChallenge || errorAuthenticate ? (
-        <div className="flex items-center space-x-1 font-bold text-red-500">
-          <XCircleIcon className="size-5" />
-          <div>{Errors.SomethingWentWrong}</div>
-        </div>
-      ) : null}
-    </div>
-  ) : (
-    <WalletSelector setHasConnected={setHasConnected} />
-  );
-};
-
-export default Signup;
