+++ apps/web/src/components/Shared/Auth/index.tsx
@@ -12,20 +12,13 @@ const Auth: FC = () => {
   const authModalType = useGlobalModalStateStore(
     (state) => state.authModalType
   );
+  console.log(authModalType);
 
   return (
     <div className="p-5">
       {authModalType === 'signup' ? (
         <div className="space-y-5">
-          {hasConnected ? (
-            <div className="space-y-1">
-              <div className="text-xl font-bold">Please sign the message.</div>
-              <div className="ld-text-gray-500 text-sm">
-                {APP_NAME} uses this signature to verify that you're the owner
-                of this address.
-              </div>
-            </div>
-          ) : (
+          {!hasConnected && (
             <div className="space-y-1">
               <div className="text-xl font-bold">Connect your wallet.</div>
               <div className="ld-text-gray-500 text-sm">
